README
Target Codevita 2020 : Contest 16

July 17 2020 11:00 AM to July 20 2020 11:00 AM

Description:

These are solutions to the coding problems of the second challenge. Problems are taken from Hackerrank and TCS Codevita challenge. All the codes are compiled at Hackerrank editor and coded in the form of functions. Most of the solutions are built in an easy manner, so that new learners can also understand the code. Codes are well commented.

The link to the problems are given below.

Happy Coding.

https://www.hackerrank.com/challenges/jumping-on-the-clouds-revisited

https://www.hackerrank.com/challenges/halloween-sale

https://www.hackerrank.com/challenges/p1-paper-cutting